+++ ComputationalMath/src/Geometry/Delaunay.h
@@ -27,11 +27,11 @@ namespace Geometry
 
         Delaunay(Triangle boundingTriangle, int nVertexCapacity);
         
-        size_t FindElement(Vertex2F point) const;
+        unsigned int FindElement(Vertex2F point) const;
 
         void FlipTest(int edgeIndex);
         
-        inline bool InCircle(size_t ai, size_t bi, size_t ci, size_t di) const;
+        inline bool InCircle(unsigned int ai, unsigned int bi, unsigned int ci, unsigned int di) const;
         
         static inline float InCircleDeterminant(Vertex2F a, Vertex2F b, Vertex2F c, Vertex2F d);
     private:
